Incorporated status effect checking for behavior, to determine a generic enemy's commands. Moved runDec and decelerate to the playeractionengine, since generic action engine doesn't use it... Made an enemy action engine, just for defining max speed and run acceleration... Maybe make this data based later? Changed some ints to floats in x and y defintions
Changed status.displace to take in a float disp and a direction XorY
Changed attemptDisplace to take the same kinds of arguments, and replaced it with a bisection type algorithm to allow for subpixel displacements.

TO DO: Finish attemptDisplace command to being like bisection. Change displace to being direction dependent.